requirements.txt
# KOICA 사업 예비조사 심사 시스템 - 의존성
# v3.1 - 개선 및 리팩토링 버전

# Web Framework
streamlit>=1.31.0,<2.0.0

# PDF Processing
PyPDF2>=3.0.1,<4.0.0

# AI & Machine Learning
google-generativeai>=0.8.0,<1.0.0
numpy>=1.24.0,<2.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0